+++ spec/hatchet/ruby_spec.rb
@@ -206,3 +206,25 @@ describe "Rack" do
     end
   end
 end
+
+describe "WEB_CONCURRENCY.sh" do
+  it "from a preceding buildpack is overwritten by this buildpack" do
+    buildpacks = [
+      "heroku/nodejs",
+      :default
+    ]
+    before_deploy = -> { run!(%Q{echo "{}" > package.json}) }
+    Hatchet::Runner.new('default_ruby', stack: DEFAULT_STACK, buildpacks: buildpacks, before_deploy: before_deploy).deploy do |app|
+      expect(app.run("cat .profile.d/WEB_CONCURRENCY.sh").strip).to be_empty
+      expect(app.run("echo $WEB_CONCURRENCY").strip).to be_empty
+      expect(app.run("echo $WEB_CONCURRENCY", :heroku => {:env => "WEB_CONCURRENCY=0"}).strip).to eq("0")
+    end
+  end
+
+  it "has defaults set with SENSIBLE_DEFAULTS on" do
+    config = { "SENSIBLE_DEFAULTS" => "1"}
+    Hatchet::Runner.new('default_ruby', stack: DEFAULT_STACK, config: config).deploy do |app|
+      expect(app.run("env")).to match("WEB_CONCURRENCY=")
+    end
+  end
+end