+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Main (unreleased)
 
+* Fix interoperability with other Heroku buildpacks' `$WEB_CONCURRENCY` handling (https://github.com/heroku/heroku-buildpack-ruby/pull/1188)
+
 ## v228 (6/24/2021)
 
 * Bundler 2.x is now 2.2.21 (https://github.com/heroku/heroku-buildpack-ruby/pull/1170)